+++ src/ParameterAccess.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Bakame\Http\StructuredFields;
 
-/**
- * @property-read MemberOrderedMap<string, Item> $parameters
- */
 interface ParameterAccess
 {
+    public function parameters(): Parameters;
+
+    public function withParameters(Parameters $parameters): static;
 }